requirements.txt
cffi==1.15.1
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
kiwisolver==1.4.4
matplotlib==3.6.1
numpy==1.23.4
packaging==21.3
Pillow==9.2.0
pycparser==2.21
pyparsing==3.0.9
PySoundFile==0.9.0.post1
python-dateutil==2.8.2
scipy==1.9.3
six==1.16.0
